Fixed a race condition when printing results
In the case where tcpping exits with a SIGINT, the parent process waits on the child process to print its status and return.  However, when returning normally (say, -c limit is reached) the SIGINT waitpid() isn't called.  Therefore, the parent often exits before the client.

This patch issues a SIGINT to the parent process when the value count is exhausted to exercise the "normal" code path.