Moved libnet calls to be global: partial migration
Moved the libnet calls to a global variable, only call the device finder from the child as it wrecks the stack and removing a packet off of the stack isn't really covered in the sample docs or anything.

There are now some weird timing problems and whatnot that got introduced somewhere along the line (though it doesn't seem to segfault anymore...) that I still need to work through.  As such, I wouldn't use this for any real use yet.