All persistent state for packets; finished port
There seem to be problems adding protocol blocks endlessly in addition to creating / deleting the libnet object.  Apparently there's some weird state that's held there.  However, since this is The Way That Things Are Going To Be With Libnet(tm) from now on, we can work around it.

Instead of creating / destroying packets after write, we now just modify the existing packets.  This could be done a lot more cleanly (by just heading into the packet buffer and rewriting it, then sending again) but I don't think it really matters for our uses.